+++ .eslintignore
@@ -1,4 +1,3 @@
 node_modules
 bin
 dist
-__testfixtures__/
