--- src/transformers/__testfixtures__/chai-should.input.js
@@ -1,21 +1,0 @@
-var expect = require('chai').expect;
-var should = require('chai').should();
-
-describe('Instantiating TextField', () => {
-    it('should set the placeholder correctly', () => {
-        textField.props.placeholder.should.equal(PLACEHOLDER);
-        textField.props.placeholder.should.not.equal(PLACEHOLDER);
-    });
-
-    it('should inherit id prop', () => {
-        dropdown.props.id.should.equal(STANDARD_PROPS.id);
-        dropdown.props.id.should.not.equal(STANDARD_PROPS.id);
-    });
-
-    it('should map open prop to visible prop', () => {
-        dropdown.props.visible.should.Throw(STANDARD_PROPS.open);
-        dropdown.props.id.should.not.Throw(STANDARD_PROPS.id);
-    });
-
-    thing1.equal(thing2);
-});
