--- src/transformers/__testfixtures__/chai-should/a-an.input.js
@@ -1,10 +1,0 @@
-expect('test').to.be.a('string', 'error message');
-expect({ foo: 'bar' }).to.be.an('object');
-expect(null).to.be.a('null');
-expect(undefined).to.be.an('undefined');
-expect(new Error()).to.be.an('error');
-expect(new Promise()).to.be.a('promise');
-expect(new Float32Array()).to.be.a('float32array');
-expect(Symbol()).to.be.a('symbol');
-
-'test'.should.be.a('string');
