--- src/transformers/__testfixtures__/chai-should/above.input.js
@@ -1,1 +1,0 @@
-expect(10).to.be.above(5);
