--- src/transformers/__testfixtures__/chai-should/above.output.js
@@ -1,1 +1,0 @@
-expect(10).toBeGreaterThan(5);
