--- src/transformers/__testfixtures__/chai-should/below.input.js
@@ -1,2 +1,0 @@
-expect(5).to.be.below(10);
-expect(5).to.be.below(10, 'error message');
