--- src/transformers/__testfixtures__/chai-should/below.output.js
@@ -1,2 +1,0 @@
-expect(5).toBeLessThan(10);
-expect(5).toBeLessThan(10);
