--- src/transformers/__testfixtures__/chai-should/eql.input.js
@@ -1,3 +1,0 @@
-expect({ foo: 'bar' }).to.eql({ foo: 'bar' });
-expect([1, 2, 3]).to.eql([1, 2, 3]);
-a.should.eql(a);
