--- src/transformers/__testfixtures__/chai-should/eql.output.js
@@ -1,3 +1,0 @@
-expect({ foo: 'bar' }).toEqual({ foo: 'bar' });
-expect([1, 2, 3]).toEqual([1, 2, 3]);
-expect(a).toEqual(a);
