--- src/transformers/__testfixtures__/chai-should/exist-defined.input.js
@@ -1,10 +1,0 @@
-expect(foo).to.exist;
-expect(bar).to.not.exist;
-expect(baz).to.not.exist;
-expect(input).exist;
-
-expect(foo).to.be.defined;
-expect(foo).not.to.be.defined;
-expect(foo).to.not.be.defined;
-
-should.exist('');
