--- src/transformers/__testfixtures__/chai-should/exist-defined.output.js
@@ -1,10 +1,0 @@
-expect(foo).toBeDefined();
-expect(bar).toBeFalsy();
-expect(baz).toBeFalsy();
-expect(input).toBeDefined();
-
-expect(foo).toBeDefined();
-expect(foo).not.toBeDefined();
-expect(foo).toBeFalsy();
-
-expect('').toBeDefined();
