--- src/transformers/__testfixtures__/chai-should/false.input.js
@@ -1,2 +1,0 @@
-expect(false).to.be.false;
-expect(0).to.not.be.false;
