--- src/transformers/__testfixtures__/chai-should/false.output.js
@@ -1,2 +1,0 @@
-expect(false).toBe(false);
-expect(0).not.toBe(false);
