--- src/transformers/__testfixtures__/chai-should/include-contain.input.js
@@ -1,4 +1,0 @@
-expect('foobar').to.have.string('bar');
-expect([1, 2, 3]).to.include(2);
-expect('foobar').to.contain('foo');
-expect({ foo: 1, bar: 2 }).to.contain({ bar: 2 });
