--- src/transformers/__testfixtures__/chai-should/include-contain.output.js
@@ -1,4 +1,0 @@
-expect('foobar').toContain('bar');
-expect([1, 2, 3]).toContain(2);
-expect('foobar').toContain('foo');
-expect({ foo: 1, bar: 2 }).toMatchObject({ bar: 2 });
