--- src/transformers/__testfixtures__/chai-should/instanceof.input.js
@@ -1,2 +1,0 @@
-expect(foo).to.be.an.instanceof(Foo);
-expect(foo).not.to.be.an.instanceof(Foo);
