--- src/transformers/__testfixtures__/chai-should/instanceof.output.js
@@ -1,2 +1,0 @@
-expect(foo).toBeInstanceOf(Foo);
-expect(foo).not.toBeInstanceOf(Foo);
