--- src/transformers/__testfixtures__/chai-should/keys.output.js
@@ -1,4 +1,0 @@
-expect([1, 2, 3]).toEqual(expect.arrayContaining([1, 2]));
-expect(Object.keys({ foo: 1, bar: 2 })).toEqual(expect.arrayContaining(Object.keys({ bar: 6, foo: 7 })));
-expect(Object.keys({ foo: 1, bar: 2, baz: 3 })).toEqual(expect.arrayContaining(['bar', 'foo']));
-expect(Object.keys({ foo: 1, bar: 2, baz: 3 })).toEqual(expect.arrayContaining(Object.keys({ bar: 6 })));
