--- src/transformers/__testfixtures__/chai-should/least.input.js
@@ -1,1 +1,0 @@
-expect(10).to.be.at.least(10);
