--- src/transformers/__testfixtures__/chai-should/least.output.js
@@ -1,1 +1,0 @@
-expect(10).toBeGreaterThanOrEqual(10);
