--- src/transformers/__testfixtures__/chai-should/lengthof.input.js
@@ -1,3 +1,0 @@
-expect([1, 2, 3]).to.have.lengthOf(3);
-expect('foobar').to.have.lengthOf(6);
-'test'.should.have.lengthOf(4);
