--- src/transformers/__testfixtures__/chai-should/lengthof.output.js
@@ -1,3 +1,0 @@
-expect([1, 2, 3]).toHaveLength(3);
-expect('foobar').toHaveLength(6);
-expect('test').toHaveLength(4);
