--- src/transformers/__testfixtures__/chai-should/match.input.js
@@ -1,1 +1,0 @@
-expect('foobar').to.match(/^foo/);
