--- src/transformers/__testfixtures__/chai-should/match.output.js
@@ -1,1 +1,0 @@
-expect('foobar').toMatch(/^foo/);
