--- src/transformers/__testfixtures__/chai-should/members.input.js
@@ -1,9 +1,0 @@
-expect([1, 2, 3]).to.include.members([3, 2]);
-expect([1, 2, 3]).to.not.include.members([3, 2, 8]);
-
-expect([4, 2]).to.have.members([2, 4], 'error message');
-expect([5, 2]).to.not.have.members([5, 2, 1]);
-
-expect([{ id: 1 }]).to.deep.include.members([{ id: 1 }]);
-
-expect({ id: 1 }).to.include.members({ id: 1 });
