--- src/transformers/__testfixtures__/chai-should/most.input.js
@@ -1,1 +1,0 @@
-expect(5).to.be.at.most(5);
