--- src/transformers/__testfixtures__/chai-should/most.output.js
@@ -1,1 +1,0 @@
-expect(5).toBeLessThanOrEqual(5);
