--- src/transformers/__testfixtures__/chai-should/nan.input.js
@@ -1,2 +1,0 @@
-expect('foo').to.be.NaN;
-expect(4).not.to.be.NaN;
