--- src/transformers/__testfixtures__/chai-should/nan.output.js
@@ -1,2 +1,0 @@
-expect('foo').toBeNaN();
-expect(4).not.toBeNaN();
