--- src/transformers/__testfixtures__/chai-should/null.input.js
@@ -1,2 +1,0 @@
-expect(null).to.be.null;
-expect(undefined).to.not.be.null;
