--- src/transformers/__testfixtures__/chai-should/null.output.js
@@ -1,2 +1,0 @@
-expect(null).toBeNull();
-expect(undefined).not.toBeNull();
