--- src/transformers/__testfixtures__/chai-should/ok.input.js
@@ -1,5 +1,0 @@
-expect('everything').to.be.ok;
-expect(1).to.be.ok;
-expect(false).to.not.be.ok;
-expect(undefined).to.not.be.ok;
-expect(null).to.not.be.ok;
