--- src/transformers/__testfixtures__/chai-should/ok.output.js
@@ -1,5 +1,0 @@
-expect('everything').toBeTruthy();
-expect(1).toBeTruthy();
-expect(false).toBeFalsy();
-expect(undefined).toBeFalsy();
-expect(null).toBeFalsy();
