--- src/transformers/__testfixtures__/chai-should/ownproperty.input.js
@@ -1,1 +1,0 @@
-expect('test').to.have.ownProperty('length');
