--- src/transformers/__testfixtures__/chai-should/ownproperty.output.js
@@ -1,1 +1,0 @@
-expect('test'.hasOwnProperty('length')).toBeTruthy();
