--- src/transformers/__testfixtures__/chai-should/ownpropertydescriptor.input.js
@@ -1,3 +1,0 @@
-expect('test').to.have.ownPropertyDescriptor('length');
-expect('test').to.have.ownPropertyDescriptor('length', { enumerable: false, configurable: false, writable: false, value: 4 });
-expect('test').not.to.have.ownPropertyDescriptor('length', { enumerable: false, configurable: false, writable: false, value: 3 });
