--- src/transformers/__testfixtures__/chai-should/ownpropertydescriptor.output.js
@@ -1,3 +1,0 @@
-expect(Object.getOwnPropertyDescriptor('test', 'length')).not.toBeUndefined();
-expect(Object.getOwnPropertyDescriptor('test', 'length')).toEqual({ enumerable: false, configurable: false, writable: false, value: 4 });
-expect(Object.getOwnPropertyDescriptor('test', 'length')).toEqual({ enumerable: false, configurable: false, writable: false, value: 3 });
