--- src/transformers/__testfixtures__/chai-should/throw.input.js
@@ -1,8 +1,0 @@
-const err = new ReferenceError('This is a bad function.');
-const fn = function() { throw err; };
-expect(fn).to.throw(ReferenceError);
-expect(fn).to.throw(Error);
-expect(fn).to.throw(/bad function/);
-expect(fn).to.not.throw('good function');
-expect(fn).to.throw(ReferenceError, /bad function/);
-expect(fn).to.throw(err);
