--- src/transformers/__testfixtures__/chai-should/throw.output.js
@@ -1,8 +1,0 @@
-const err = new ReferenceError('This is a bad function.');
-const fn = function() { throw err; };
-expect(fn).toThrowError(ReferenceError);
-expect(fn).toThrowError(Error);
-expect(fn).toThrowError(/bad function/);
-expect(fn).not.toThrowError('good function');
-expect(fn).toThrowError(ReferenceError);
-expect(fn).toThrowError(err);
