--- src/transformers/__testfixtures__/chai-should/true.input.js
@@ -1,2 +1,0 @@
-expect(true).to.be.true;
-expect(1).to.not.be.true;
