--- src/transformers/__testfixtures__/chai-should/true.output.js
@@ -1,2 +1,0 @@
-expect(true).toBe(true);
-expect(1).not.toBe(true);
