--- src/transformers/__testfixtures__/chai-should/undefined.input.js
@@ -1,2 +1,0 @@
-expect(undefined).to.be.undefined;
-expect(null).to.not.be.undefined;
