--- src/transformers/__testfixtures__/chai-should/undefined.output.js
@@ -1,2 +1,0 @@
-expect(undefined).toBeUndefined();
-expect(null).toBeDefined();
