--- src/transformers/__testfixtures__/chai-should/within.output.js
@@ -1,16 +1,0 @@
-expect(7).toBeLessThanOrEqual(5);
-expect(7).toBeGreaterThanOrEqual(10);
-
-expect('foo'.length).toBeLessThanOrEqual(2);
-
-expect('foo'.length).toBeGreaterThanOrEqual(4);
-expect([1, 2, 3].length).toBeLessThanOrEqual(2);
-expect([1, 2, 3].length).toBeGreaterThanOrEqual(4);
-
-expect('foo'.length).toBeLessThanOrEqual(2);
-
-expect('foo'.length).toBeGreaterThanOrEqual(4);
-
-expect(5).toBeLessThanOrEqual(2);
-
-expect(5).toBeGreaterThanOrEqual(4);
