+++ src/transformers/chai-should.test.js
@@ -1,7 +1,5 @@
 /* eslint-env jest */
-import path from 'path';
 import chalk from 'chalk';
-import { defineTest } from 'jscodeshift/dist/testUtils';
 import { wrapPlugin } from '../utils/test-helpers';
 import plugin from './chai-should';
 
@@ -13,39 +11,6 @@ beforeEach(() => {
     consoleWarnings = [];
     console.warn = v => consoleWarnings.push(v);
 });
-
-const transformersPath = path.join(__dirname, 'transformers');
-
-const createTest = name => {
-    defineTest(transformersPath, 'chai-should', null, path.join('chai-should', name));
-};
-
-defineTest(transformersPath, 'chai-should', null, 'chai-should');
-
-createTest('a-an');
-createTest('above');
-createTest('below');
-createTest('eql');
-createTest('equal');
-createTest('exist-defined');
-createTest('false');
-createTest('include-contain');
-createTest('instanceof');
-createTest('keys');
-createTest('least');
-createTest('lengthof');
-createTest('match');
-createTest('members');
-createTest('most');
-createTest('nan');
-createTest('null');
-createTest('ok');
-createTest('ownproperty');
-createTest('ownpropertydescriptor');
-createTest('throw');
-createTest('true');
-createTest('undefined');
-createTest('within');
 
 function testChanged(_msg, _source, _expectedOutput) {
     let msg = _msg;
@@ -64,6 +29,437 @@ function testChanged(_msg, _source, _expectedOutput) {
         expect(consoleWarnings).toEqual([]);
     });
 }
+
+testChanged(
+    'removes imports and does basic conversions of should and expect',
+    `
+        var expect = require('chai').expect;
+        var should = require('chai').should();
+
+        describe('Instantiating TextField', () => {
+            it('should set the placeholder correctly', () => {
+                textField.props.placeholder.should.equal(PLACEHOLDER);
+                textField.props.placeholder.should.not.equal(PLACEHOLDER);
+            });
+
+            it('should inherit id prop', () => {
+                dropdown.props.id.should.equal(STANDARD_PROPS.id);
+                dropdown.props.id.should.not.equal(STANDARD_PROPS.id);
+            });
+
+            it('should map open prop to visible prop', () => {
+                dropdown.props.visible.should.Throw(STANDARD_PROPS.open);
+                dropdown.props.id.should.not.Throw(STANDARD_PROPS.id);
+            });
+
+            thing1.equal(thing2);
+        });
+    `,
+    `
+        describe('Instantiating TextField', () => {
+            it('should set the placeholder correctly', () => {
+                expect(textField.props.placeholder).toBe(PLACEHOLDER);
+                expect(textField.props.placeholder).not.toBe(PLACEHOLDER);
+            });
+
+            it('should inherit id prop', () => {
+                expect(dropdown.props.id).toBe(STANDARD_PROPS.id);
+                expect(dropdown.props.id).not.toBe(STANDARD_PROPS.id);
+            });
+
+            it('should map open prop to visible prop', () => {
+                expect(dropdown.props.visible).toThrowError(STANDARD_PROPS.open);
+                expect(dropdown.props.id).not.toThrowError(STANDARD_PROPS.id);
+            });
+
+            thing1.equal(thing2);
+        });
+    `
+);
+
+testChanged(
+    'converts "a-an"',
+    `
+        expect('test').to.be.a('string', 'error message');
+        expect({ foo: 'bar' }).to.be.an('object');
+        expect(null).to.be.a('null');
+        expect(undefined).to.be.an('undefined');
+        expect(new Error()).to.be.an('error');
+        expect(new Promise()).to.be.a('promise');
+        expect(new Float32Array()).to.be.a('float32array');
+        expect(Symbol()).to.be.a('symbol');
+
+        'test'.should.be.a('string');
+    `,
+    `
+        expect(typeof 'test').toBe('string');
+        expect(typeof { foo: 'bar' }).toBe('object');
+        expect(null).toBeNull();
+        expect(undefined).toBeUndefined();
+        expect(typeof new Error()).toBe('error');
+        expect(typeof new Promise()).toBe('promise');
+        expect(typeof new Float32Array()).toBe('float32array');
+        expect(typeof Symbol()).toBe('symbol');
+
+        expect(typeof 'test').toBe('string');
+    `
+);
+
+testChanged(
+    'converts "above"',
+    `
+        expect(10).to.be.above(5);
+    `,
+    `
+        expect(10).toBeGreaterThan(5);
+    `
+);
+
+testChanged(
+    'converts "below"',
+    `
+        expect(5).to.be.below(10);
+        expect(5).to.be.below(10, 'error message');
+    `,
+    `
+        expect(5).toBeLessThan(10);
+        expect(5).toBeLessThan(10);
+    `
+);
+
+testChanged(
+    'converts "eql"',
+    `
+        expect({ foo: 'bar' }).to.eql({ foo: 'bar' });
+        expect([1, 2, 3]).to.eql([1, 2, 3]);
+        a.should.eql(a);
+    `,
+    `
+        expect({ foo: 'bar' }).toEqual({ foo: 'bar' });
+        expect([1, 2, 3]).toEqual([1, 2, 3]);
+        expect(a).toEqual(a);
+    `
+);
+
+testChanged(
+    'converts "equal"',
+    `
+        expect('hello').to.equal('hello');
+        expect(42).to.equal(42);
+        expect(1).to.not.equal(true);
+        expect({ foo: 'bar' }).to.not.equal({ foo: 'bar' });
+        expect({ foo: 'bar' }).to.deep.equal({ foo: 'bar' });
+
+        should.equal('foo', 'foo');
+        should.not.equal('foo', 'bar');
+    `,
+    `
+        expect('hello').toBe('hello');
+        expect(42).toBe(42);
+        expect(1).not.toBe(true);
+        expect({ foo: 'bar' }).not.toBe({ foo: 'bar' });
+        expect({ foo: 'bar' }).toEqual({ foo: 'bar' });
+
+        expect('foo').toBe('foo');
+        expect('foo').not.toBe('bar');
+    `
+);
+
+testChanged(
+    'converts "exist-defined"',
+    `
+        expect(foo).to.exist;
+        expect(bar).to.not.exist;
+        expect(baz).to.not.exist;
+        expect(input).exist;
+
+        expect(foo).to.be.defined;
+        expect(foo).not.to.be.defined;
+        expect(foo).to.not.be.defined;
+
+        should.exist('');
+    `,
+    `
+        expect(foo).toBeDefined();
+        expect(bar).toBeFalsy();
+        expect(baz).toBeFalsy();
+        expect(input).toBeDefined();
+
+        expect(foo).toBeDefined();
+        expect(foo).not.toBeDefined();
+        expect(foo).toBeFalsy();
+
+        expect('').toBeDefined();
+    `
+);
+
+testChanged(
+    'converts "false"',
+    `
+        expect(false).to.be.false;
+        expect(0).to.not.be.false;
+    `,
+    `
+        expect(false).toBe(false);
+        expect(0).not.toBe(false);
+    `
+);
+
+testChanged(
+    'converts "include-contain"',
+    `
+        expect('foobar').to.have.string('bar');
+        expect([1, 2, 3]).to.include(2);
+        expect('foobar').to.contain('foo');
+        expect({ foo: 1, bar: 2 }).to.contain({ bar: 2 });
+    `,
+    `
+        expect('foobar').toContain('bar');
+        expect([1, 2, 3]).toContain(2);
+        expect('foobar').toContain('foo');
+        expect({ foo: 1, bar: 2 }).toMatchObject({ bar: 2 });
+    `
+);
+
+testChanged(
+    'converts "instanceof"',
+    `
+        expect(foo).to.be.an.instanceof(Foo);
+        expect(foo).not.to.be.an.instanceof(Foo);
+    `,
+    `
+        expect(foo).toBeInstanceOf(Foo);
+        expect(foo).not.toBeInstanceOf(Foo);
+    `
+);
+
+testChanged(
+    'converts "keys"',
+    `
+        expect([1, 2, 3]).to.have.all.keys(1, 2);
+        expect({ foo: 1, bar: 2 }).to.have.all.keys({ bar: 6, foo: 7 });
+        expect({ foo: 1, bar: 2, baz: 3 }).to.contain.all.keys(['bar', 'foo']);
+        expect({ foo: 1, bar: 2, baz: 3 }).to.contain.all.keys({ bar: 6 });
+    `,
+    `
+        expect([1, 2, 3]).toEqual(expect.arrayContaining([1, 2]));
+        expect(Object.keys({ foo: 1, bar: 2 })).toEqual(expect.arrayContaining(Object.keys({ bar: 6, foo: 7 })));
+        expect(Object.keys({ foo: 1, bar: 2, baz: 3 })).toEqual(expect.arrayContaining(['bar', 'foo']));
+        expect(Object.keys({ foo: 1, bar: 2, baz: 3 })).toEqual(expect.arrayContaining(Object.keys({ bar: 6 })));
+    `
+);
+
+testChanged(
+    'converts "least"',
+    `
+        expect(10).to.be.at.least(10);
+    `,
+    `
+        expect(10).toBeGreaterThanOrEqual(10);
+    `
+);
+
+testChanged(
+    'converts "lengthof"',
+    `
+        expect([1, 2, 3]).to.have.lengthOf(3);
+        expect('foobar').to.have.lengthOf(6);
+        'test'.should.have.lengthOf(4);
+    `,
+    `
+        expect([1, 2, 3]).toHaveLength(3);
+        expect('foobar').toHaveLength(6);
+        expect('test').toHaveLength(4);
+    `
+);
+
+testChanged(
+    'converts "match"',
+    `
+        expect('foobar').to.match(/^foo/);
+    `,
+    `
+        expect('foobar').toMatch(/^foo/);
+    `
+);
+
+testChanged(
+    'converts "members"',
+    `
+        expect([1, 2, 3]).to.include.members([3, 2]);
+        expect([1, 2, 3]).to.not.include.members([3, 2, 8]);
+
+        expect([4, 2]).to.have.members([2, 4], 'error message');
+        expect([5, 2]).to.not.have.members([5, 2, 1]);
+
+        expect([{ id: 1 }]).to.deep.include.members([{ id: 1 }]);
+
+        expect({ id: 1 }).to.include.members({ id: 1 });
+    `,
+    `
+        expect([1, 2, 3]).toEqual(expect.arrayContaining([3, 2]));
+        expect([1, 2, 3]).not.toEqual(expect.arrayContaining([3, 2, 8]));
+
+        expect([4, 2]).toEqual(expect.arrayContaining([2, 4]));
+        expect([5, 2]).not.toEqual(expect.arrayContaining([5, 2, 1]));
+
+        expect([{ id: 1 }]).toEqual(expect.arrayContaining([{ id: 1 }]));
+
+        expect({ id: 1 }).toEqual(expect.objectContaining({ id: 1 }));
+    `
+);
+
+testChanged(
+    'converts "most"',
+    `
+        expect(5).to.be.at.most(5);
+    `,
+    `
+        expect(5).toBeLessThanOrEqual(5);
+    `
+);
+
+testChanged(
+    'converts "nan"',
+    `
+        expect('foo').to.be.NaN;
+        expect(4).not.to.be.NaN;
+    `,
+    `
+        expect('foo').toBeNaN();
+        expect(4).not.toBeNaN();
+    `
+);
+
+testChanged(
+    'converts "null"',
+    `
+        expect(null).to.be.null;
+        expect(undefined).to.not.be.null;
+    `,
+    `
+        expect(null).toBeNull();
+        expect(undefined).not.toBeNull();
+    `
+);
+
+testChanged(
+    'converts "ok"',
+    `
+        expect('everything').to.be.ok;
+        expect(1).to.be.ok;
+        expect(false).to.not.be.ok;
+        expect(undefined).to.not.be.ok;
+        expect(null).to.not.be.ok;
+    `,
+    `
+        expect('everything').toBeTruthy();
+        expect(1).toBeTruthy();
+        expect(false).toBeFalsy();
+        expect(undefined).toBeFalsy();
+        expect(null).toBeFalsy();
+    `
+);
+
+testChanged(
+    'converts "ownproperty"',
+    `expect('test').to.have.ownProperty('length')`,
+    `expect('test'.hasOwnProperty('length')).toBeTruthy()`
+);
+
+testChanged(
+    'converts "ownpropertydescriptor"',
+    `
+        expect('test').to.have.ownPropertyDescriptor('length');
+        expect('test').to.have.ownPropertyDescriptor('length', { enumerable: false, configurable: false, writable: false, value: 4 });
+        expect('test').not.to.have.ownPropertyDescriptor('length', { enumerable: false, configurable: false, writable: false, value: 3 });
+    `,
+    `
+        expect(Object.getOwnPropertyDescriptor('test', 'length')).not.toBeUndefined();
+        expect(Object.getOwnPropertyDescriptor('test', 'length')).toEqual({ enumerable: false, configurable: false, writable: false, value: 4 });
+        expect(Object.getOwnPropertyDescriptor('test', 'length')).toEqual({ enumerable: false, configurable: false, writable: false, value: 3 });
+    `
+);
+
+testChanged(
+    'converts "throw"',
+    `
+        const err = new ReferenceError('This is a bad function.');
+        const fn = function() { throw err; };
+        expect(fn).to.throw(ReferenceError);
+        expect(fn).to.throw(Error);
+        expect(fn).to.throw(/bad function/);
+        expect(fn).to.not.throw('good function');
+        expect(fn).to.throw(ReferenceError, /bad function/);
+        expect(fn).to.throw(err);
+    `,
+    `
+        const err = new ReferenceError('This is a bad function.');
+        const fn = function() { throw err; };
+        expect(fn).toThrowError(ReferenceError);
+        expect(fn).toThrowError(Error);
+        expect(fn).toThrowError(/bad function/);
+        expect(fn).not.toThrowError('good function');
+        expect(fn).toThrowError(ReferenceError);
+        expect(fn).toThrowError(err);
+    `
+);
+
+testChanged(
+    'converts "true"',
+    `
+        expect(true).to.be.true;
+        expect(1).to.not.be.true;
+    `,
+    `
+        expect(true).toBe(true);
+        expect(1).not.toBe(true);
+    `
+);
+
+testChanged(
+    'converts "undefined"',
+    `
+        expect(undefined).to.be.undefined;
+        expect(null).to.not.be.undefined;
+    `,
+    `
+        expect(undefined).toBeUndefined();
+        expect(null).toBeDefined();
+    `
+);
+
+testChanged(
+    'converts "within"',
+    `
+        expect(7).to.be.within(5, 10);
+
+        expect('foo').to.have.length.within(2, 4);
+        expect([1, 2, 3]).to.have.length.within(2, 4);
+
+        expect('foo').to.have.length.within(2, 4, 'error message');
+
+        (5).should.be.within(2, 4);
+    `,
+    `
+        expect(7).toBeLessThanOrEqual(5);
+        expect(7).toBeGreaterThanOrEqual(10);
+
+        expect('foo'.length).toBeLessThanOrEqual(2);
+
+        expect('foo'.length).toBeGreaterThanOrEqual(4);
+        expect([1, 2, 3].length).toBeLessThanOrEqual(2);
+        expect([1, 2, 3].length).toBeGreaterThanOrEqual(4);
+
+        expect('foo'.length).toBeLessThanOrEqual(2);
+
+        expect('foo'.length).toBeGreaterThanOrEqual(4);
+
+        expect(5).toBeLessThanOrEqual(2);
+
+        expect(5).toBeGreaterThanOrEqual(4);
+    `
+);
 
 testChanged('expect("123").to.eql("123");', 'expect("123").toEqual("123");');
